CHG: Compatibility with Ubuntu 11.10
git-svn-id: https://elves.srl.caltech.edu/svn/nusim/trunk@425 f61b1093-7add-4ab6-ac1b-eb05caf01ebb

+++ src/modules/inc/NGUIOptionsSinglePointing.h
@@ -46,7 +46,7 @@ class NGUIOptionsSinglePointing : public TGCompositeFrame
   // public interface:
  public:
   //! Standard constructor
-  NGUIOptionsSinglePointing(const TGWindow* Parent, NPointing* Pointing, unsigned int ID); 
+  NGUIOptionsSinglePointing(const TGWindow* Parent, NPointing* Pointing, unsigned long ID); 
   //! Default destructor
   virtual ~NGUIOptionsSinglePointing();
 
@@ -74,7 +74,7 @@ class NGUIOptionsSinglePointing : public TGCompositeFrame
   //! The SinglePointing itself
   NPointing* m_Pointing;
   //! ID of this pointing
-  unsigned int m_ID;
+  unsigned long m_ID;
   
   MGUIEEntry* m_Dec;
   MGUIEEntry* m_Ra;
